Fix bug whereby not providing initialSuggestions to LinkControl disabled the ability to search at all
Test failures led to uncovering of a bug whereby the URLInput component’s `updateSuggestions` method would exit early it initialSuggestions wasn’t provided. As initialSuggestions is an optional prop this means that all fetching of search suggestions was disabled if the initialSuggestions wasn’t provided.

Fixed this bug and all tests by improving the conditionals.