Refactor to remove seperate initialSuggestions query
As per the GH thread below, this refactors the code to avoid the need to introduce a new fetch API around `__experimentalInitialSuggestions`. Now instead we simply reuse the existing fetchSuggestions handler to get the initial results. The only different being we introduce an arguments object to queries to restrict the number of results displayed for initial Suggestions.

See https://github.com/WordPress/gutenberg/pull/19458#issuecomment-573151163